+++ packages/frontend/src/components/common/Filters/FilterInputs/utils.ts
@@ -81,7 +81,9 @@ export const getFilterOperatorOptions = (
                 FilterOperator.EQUALS,
                 FilterOperator.NOT_EQUALS,
                 FilterOperator.LESS_THAN,
+                FilterOperator.LESS_THAN_OR_EQUAL,
                 FilterOperator.GREATER_THAN,
+                FilterOperator.GREATER_THAN_OR_EQUAL,
                 FilterOperator.IN_BETWEEN,
                 FilterOperator.NOT_IN_BETWEEN,
             ]);